APR-550 tech debt: remove UKRLP API, this has moved to RoATP service
