+++ src/SFA.DAS.ApplyService.InternalApi/Infrastructure/IRoatpApiClient.cs
@@ -4,6 +4,7 @@
     using System.Collections.Generic;
     using System.Threading.Tasks;
     using Domain.Roatp;
+    using Models.Ukrlp;
     using SFA.DAS.ApplyService.InternalApi.Models.Roatp;
 
     public interface IRoatpApiClient
@@ -11,5 +12,6 @@
         Task<IEnumerable<ProviderType>> GetProviderTypes();
         Task<DuplicateCheckResponse> DuplicateUKPRNCheck(Guid organisationId, long ukprn);
         Task<OrganisationRegisterStatus> GetOrganisationRegisterStatus(Guid organisationId);
+        Task<UkprnLookupResponse> GetUkrlpDetails(string ukprn);
     }
 }
