--- src/SFA.DAS.ApplyService.InternalApi/Infrastructure/IUkrlpApiClient.cs
@@ -1,10 +1,0 @@
-﻿namespace SFA.DAS.ApplyService.InternalApi.Infrastructure
-{
-    using System.Threading.Tasks;
-    using Models.Ukrlp;
-
-    public interface IUkrlpApiClient
-    {
-        Task<UkprnLookupResponse> GetTrainingProviderByUkprn(long ukprn);
-    }
-}
